feat:  refactor input
